+++ packages/react-widgets/src/widgets/BarWidget.js
@@ -164,6 +164,7 @@ function BarWidget({
             height={height}
             animation={animation}
             filterable={filterable}
+            isLoading={isLoading}
           />
         )}
       </WidgetWithAlert>